refactor: use same file for batch operation hooks
