--- operate/client/src/modules/hooks/useDeleteBatchOperationMutationRequestBody.ts
@@ -1,42 +1,0 @@
-/*
- * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH under
- * one or more contributor license agreements. See the NOTICE file distributed
- * with this work for additional information regarding copyright ownership.
- * Licensed under the Camunda License 1.0. You may not use this file
- * except in compliance with the Camunda License 1.0.
- */
-
-import {useSearchParams} from 'react-router-dom';
-import {variableFilterStore} from 'modules/stores/variableFilter';
-import {processInstancesSelectionStore} from 'modules/stores/processInstancesSelection';
-import {buildMutationRequestBody} from 'modules/utils/buildMutationRequestBody';
-
-/**
- * Hook for building the request body for delete batch operations.
- * Unlike running operations (cancel, retry), delete operations only work on
- * finished instances (COMPLETED or TERMINATED).
- */
-const useDeleteBatchOperationMutationRequestBody = () => {
-  const variable = variableFilterStore.variable;
-  const [searchParams] = useSearchParams();
-
-  const {
-    selectedProcessInstanceIds,
-    excludedProcessInstanceIds,
-    checkedFinishedProcessInstanceIds,
-  } = processInstancesSelectionStore;
-
-  const includeIds =
-    selectedProcessInstanceIds.length > 0
-      ? checkedFinishedProcessInstanceIds
-      : [];
-
-  return buildMutationRequestBody({
-    searchParams,
-    includeIds,
-    excludeIds: excludedProcessInstanceIds,
-    variableFilter: variable,
-  });
-};
-
-export {useDeleteBatchOperationMutationRequestBody};
